MAJOR API CHANGE: - change matcher #= to #equal: e.g. "1 should = 1" would now be written "1 should equal: 1" - update all code to use new API
Motivation: the #= magic made it impossible to store matchers in dictionaries, and #hash was implemented to signal an error explaining as much. Unfortunately, #= and #hash are deeply ingrained Smalltalk concepts, and are assumed to work as expected. In Pharo 3.0, the debugger tried to put matchers in a dictionary, causing an infinite error loop whenever a matcher failed.

See http://forum.world.st/Phexample-API-Change-Proposal-was-Phexample-Image-Destroying-Bug-td4755787.html for full discussion. Also discussed on Squeak Dev with no opposition